qc/modules/fastqc: consolidate QC module tests for 'fastqc' module.
